Ensure all ENV vars are `String`s.
Otherwise `Mixlib::Shellout` fails in weird ways when
`remote_exec = false` because the `state` has a port which is a numeric,
and `ENV` only likes `String`s.

Also use `shellout` for all commands so the serverspec run gets the state -
necessary to pick up the hostname, port, username, and SSH key so we can SSH
to the instance for the tests.